Move Backend filters into domain
